requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
# -f https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.8.0-py3-none-any.whl
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cssselect==1.1.0
Cython==0.29.24
Django==3.2.9
django-background-tasks==1.2.5
django-compat==1.0.15
django-crontab==0.7.1
djangorestframework==3.12.4
djangorestframework-dataclasses==1.0.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
flatbuffers==2.0
gast==0.4.0
gensim==4.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
hdbscan==0.8.27
hnswlib==0.5.2
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.8.2
jieba3k==0.35.1
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
llvmlite==0.37.0
logtail-python==0.1.2
lxml==4.6.4
Markdown==3.3.6
msgpack==1.0.3
mysqlclient==2.1.0
marshmallow==3.14.1
desert==2020.11.18
mypy-extensions==0.4.3
typing-inspect==0.7.1
newspaper3k==0.2.8
nltk==3.6.5
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynndescent==0.5.5
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
schedule==1.1.0
scikit-learn==1.0.1
scipy==1.7.3
sentence-transformers==2.1.0
sentencepiece==0.1.96
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
sqlparse==0.4.2
tensorflow-cpu
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-text==2.7.3
termcolor==1.1.0
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.10.3
torch==1.10.0+cpu
torchvision==0.11.1
tqdm==4.62.3
transformers==4.12.5
typing_extensions==4.0.1
umap-learn==0.5.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0

# absl-py==1.0.0
# astunparse==1.6.3
# attrs==21.2.0
# awsebcli==3.18.1
# bcrypt==3.2.0
# beautifulsoup4==4.10.0
# blessed==1.19.0
# botocore==1.15.49
# cached-property==1.5.2
# cachetools==4.2.4
# cement==2.8.2
# certifi==2021.10.8
# cffi==1.15.0
# chardet==3.0.4
# charset-normalizer==2.0.7
# click==8.0.3
# colorama==0.4.4
# cryptography==36.0.0
# cssselect==1.1.0
# cycler==0.11.0
# Cython==0.29.24
# Django==2.2
# django-background-tasks==1.2.5
# django-compat==1.0.15
# django-crontab==0.7.1
# djangorestframework==3.12.4
# docker==4.4.4
# docker-compose==1.25.5
# dockerpty==0.4.1
# docopt==0.6.2
# docutils==0.15.2
# feedfinder2==0.0.4
# feedparser==6.0.8
# filelock==3.3.2
# flatbuffers==2.0
# fonttools==4.28.2
# future==0.16.0
# gast==0.4.0
# gensim==3.8.3
# google-auth==2.3.3
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.42.0
# h5py==3.6.0
# hdbscan==0.8.27
# hnswlib==0.5.2
# huggingface-hub==0.1.2
# idna==2.7
# importlib-metadata==4.8.2
# jieba3k==0.35.1
# jmespath==0.10.0
# joblib==1.1.0
# jsonschema==3.2.0
# keras==2.7.0
# Keras-Preprocessing==1.1.2
# kiwisolver==1.3.2
# libclang==12.0.0
# llvmlite==0.37.0
# logtail-python==0.1.2
# lxml==4.6.4
# Markdown==3.3.6
# matplotlib==3.5.0
# msgpack==1.0.2
# mysqlclient==2.0.3
# newspaper3k==0.2.8
# nltk==3.6.5
# numba==0.54.1
# numpy==1.20.3
# oauthlib==3.1.1
# opt-einsum==3.3.0
# packaging==21.3
# pandas==1.3.4
# paramiko==2.8.0
# pathspec==0.5.9
# Pillow==8.4.0
# protobuf==3.19.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.21
# PyNaCl==1.4.0
# pynndescent==0.5.5
# pyparsing==3.0.6
# pyrsistent==0.18.0
# python-dateutil==2.8.0
# pytz==2021.3
# PyYAML==5.3.1
# regex==2021.11.2
# requests==2.26.0
# requests-file==1.5.1
# requests-oauthlib==1.3.0
# rsa==4.8
# sacremoses==0.0.46
# schedule==1.1.0
# scikit-learn==1.0.1
# scipy==1.7.3
# semantic-version==2.5.0
# sentence-transformers==2.1.0
# sentencepiece==0.1.96
# setuptools-scm==6.3.2
# sgmllib3k==1.0.0
# six
# smart-open==5.2.1
# soupsieve==2.3
# sqlparse==0.4.2
# # tensorboard==2.7.0
# # tensorboard-data-server==0.6.1
# # tensorboard-plugin-wit==1.8.0
# # tensorflow==2.7.0
# # tensorflow-estimator==2.7.0
# # tensorflow-hub==0.12.0
# # tensorflow-io-gcs-filesystem==0.22.0
# # tensorflow-text==2.7.3
# termcolor==1.1.0
# texttable==1.6.4
# threadpoolctl==3.0.0
# tinysegmenter==0.3
# tldextract==3.1.2
# tokenizers==0.10.3
# tomli==1.2.2
# top2vec==1.0.26
# torch==1.6.0+cpu
# torchvision==0.11.1
# tqdm==4.62.3
# transformers==4.12.5
# typing_extensions==4.0.0
# umap==0.1.1
# umap-learn==0.5.2
# urllib3==1.24.3
# wcwidth==0.1.9
# websocket-client==0.59.0
# Werkzeug==2.0.2
# wordcloud==1.8.1
# wrapt==1.13.3
# zipp==3.6.0

# asgiref==3.2.7
# Django==2.2
# pytz==2021.3
# sqlparse==0.4.2
# gunicorn==20.0.4
# djangorestframework==3.12.4
# schedule==1.1.0
# mysqlclient
# django_crontab
# newspaper3k
# feedparser
# logtail-python==0.1.2
# msgpack==1.0.2
# django-background-tasks==1.2.5
# django-compat==1.0.15
# awsebcli==3.18.1
# myproject
# pandas==1.3.4
# numpy
# gensim==4.1.2
# scipy==1.7.3
# smart-open==5.2.1
# umap==0.1.1
# cython==0.29.24
# hdbscan==0.8.27
# scikit-learn==1.0.1
# threadpoolctl==3.0.0
# hnswlib==0.5.2
# huggingface-hub==0.1.2
# packaging==21.3
# pyparsing==3.0.6
# sacremoses==0.0.46
# sentence-transformers==2.1.0
# sentencepiece==0.1.96
# tokenizers==0.10.3
# # torchvision==0.11.1
# torch==1.6.0+cpu
# transformers==4.12.5
# typing-extensions==4.0.0
# absl-py==1.0.0
# astunparse==1.6.3
# cachetools==4.2.4
# cycler==0.11.0
# flatbuffers==2.0
# fonttools==4.28.2
# gast==0.4.0
# google-auth==2.3.3
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.42.0
# h5py==3.6.0
# importlib-metadata==4.8.2
# keras==2.7.0
# keras-preprocessing==1.1.2
# kiwisolver==1.3.2
# libclang==12.0.0
# llvmlite==0.37.0
# markdown==3.3.6
# matplotlib==3.5.0
# numba==0.54.1
# numpy==1.20.3
# oauthlib==3.1.1
# opt-einsum==3.3.0
# protobuf==3.19.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pynndescent==0.5.5
# requests
# requests-oauthlib==1.3.0
# rsa==4.8
# setuptools-scm==6.3.2
# six
# # tensorflow==2.7.0
# # tensorflow-estimator==2.7.0
# # tensorflow-hub==0.12.0
# # tensorflow-io-gcs-filesystem==0.22.0
# # tensorflow-text==2.7.3
# tomli==1.2.2
# top2vec==1.0.26
# umap-learn==0.5.2
# werkzeug==2.0.2
# wordcloud==1.8.1
# wrapt==1.13.3
# zipp==3.6.0
